requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
Deprecated==1.2.14
dj-database-url==2.1.0
Django==3.2.22
django-admin-geomap==1.6
django-cors-headers==4.1.0
django-nested-admin==4.0.2
django-oauth-toolkit==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.6
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.9
pycparser==2.21
python-monkey-business==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.0
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==6.6.0
wrapt==1.15.0